isr: stop processing modem in interrupt handler
Disable interrupts for the modem and just check its LSR during our
activity loop.  When there's a byte available, read it and process
it.

This properly slows down input so we can handle it without dropping
things so we can see a full AT&V1 without dropping characters even
with the DLAB cranked to 14400.