Build: Automatically calculate data-loc
The segments in crt0 are not doing what I thought they were, and I
think to actually put things for the whole project in the particular
order in the final binary, I need a custom linker script which I
don't really want to deal with.

I would rather just put the data right after the code segments since
it's all in the same RAM page anyway, which would result in smaller
files being sent over the parallel interface each time, but this at
least gets things building and running reliably with data put as
high into the 16k page as possible.