Ignore power button as first key seen, make it actually power down
I wasted a whole day debugging why this was seemingly crashing when
it turns out it was just getting into process_keyboard() and
peekkey() was returning KEY_POWER all the time at init time, which
just rebooted the machine.

By assigning KEY_POWER to lastkey, the debouncer will ignore it as
the first key pressed.  Also add a blinking sequence before calling
the powerdown function instead of rebooting, to make it clear it's
not just crashing.